fix func update category

+++ src/main/java/com/SWP391/G3PCoffee/controller/ShopController.java
@@ -22,7 +22,7 @@ public class ShopController {
     private ProductRepository productRepository;
 
     @GetMapping("/shop")
-    public String showShop(@RequestParam(value = "categoryId", required = false) Integer categoryId, Model model) {
+    public String showShop(@RequestParam(value = "categoryId", required = false) Long categoryId, Model model) {
         // Lấy danh sách danh mục
         List<Category> categories = categoryRepository.findAll();
         model.addAttribute("categories", categories);
